style: Move SelectorMap to its own submodule.
